Fixed CullableOnUncontained.kt. Also, added [kill] method to IGameEntity.kt which accepts props param. In the [kill] method override in GameEntity.kt, a prop is accessed which contains the cause of death for the entity and logs it.

+++ core/src/main/com/engine/spawns/SpawnerForEvent.kt
@@ -10,7 +10,7 @@ class SpawnerForEvent(
     private val predicate: (Event) -> Boolean,
     private val spawnSupplier: () -> Spawn,
     override val eventKeyMask: ObjectSet<Any> = ObjectSet(),
-    shouldBeCulled: () -> Boolean = { false },
+    shouldBeCulled: (Float) -> Boolean = { false },
     onCull: () -> Unit = {},
 ) : Spawner(shouldBeCulled, onCull), IEventListener {
